feat: wine games updated

+++ one_key_wine_games_update.sh
@@ -109,22 +109,13 @@ do_when_old_file_exist() {
     if [ -d "$TEMPLATE_WINE_APP_PATH" ]; then
         info "[info]  Template app path:    $TEMPLATE_WINE_APP_PATH"
         info "[info]  New version app path: $new_version_app_path"
+        info "[info]  Game exe name:        $game_exe_name"
 
         info "    Copying template app to new version app"
 
         cp -R "$TEMPLATE_WINE_APP_PATH" "$new_version_app_path"
 
         OpenWineInlineConfigApp "$new_version_app_path"
-
-        sleep 80
-
-        do_when_new_file_exist "$install_folder_name" "$old_version_file_name" "$new_version_file_name" "$old_version_app_path" "$new_version_app_path"
-
-        toxicConfigApp="${new_version_app_path}/${InlineConfigAppInToxicGame}"
-        if [ -d "$toxicConfigApp" ]; then
-            write_game_exe_file_to_config_file "$install_folder_name" "$old_version_file_name" "$new_version_file_name" "$old_version_app_path" "$new_version_app_path" "$game_exe_name"
-        fi
-
     else
         echo
         error "   [error]  new version app:'${new_version_file_name}' not exist!"
@@ -139,26 +130,31 @@ write_game_exe_file_to_config_file() {
     new_version_app_path=$5
     game_exe_name=$6
 
-    sleep 5
-    killall launcher
-    sleep 10
-
+    toxicConfigApp="${new_version_app_path}/${InlineConfigAppInToxicGame}"
     toxicConfigJson="${new_version_app_path}/${InlineConfigJsonInToxicGame}"
 
-    # "exe_path" : "ToxicGame\/drive_c\/Games\/Assassins Creed IV Black Flag\/AC4BFSP.exe",
-    # "exe_path" : "ToxicGame\/drive_c\/Games\/Assassins Creed IV Black Flag\/AC4BFSP.exe",
-    # "backend" : "dxmt",
-    # "backend" : "dxmt",
+    if [ -d "$toxicConfigApp" ]; then
+        sleep 5
+        killall launcher
+        sleep 10
 
-    # check if a '$game_exe_name' is not empty
-    if [[ -n "$game_exe_name" ]]; then
-        node -e "let pkg=require('${toxicConfigJson}'); pkg['exe_path'] = \"ToxicGame\/drive_c\/Games\/${install_folder_name}\/${game_exe_name}\"; require('fs').writeFileSync('${toxicConfigJson}', JSON.stringify(pkg, null, 2));"
+        # "exe_path" : "ToxicGame\/drive_c\/Games\/Assassins Creed IV Black Flag\/AC4BFSP.exe",
+        # "exe_path" : "ToxicGame\/drive_c\/Games\/Assassins Creed IV Black Flag\/AC4BFSP.exe",
+        # "backend" : "dxmt",
+        # "backend" : "dxmt",
 
-        if [[ "$my_global_file_type" == *"mt"* ]]; then
-            node -e "let pkg=require('${toxicConfigJson}'); pkg['backend'] = 'dxmt'; require('fs').writeFileSync('${toxicConfigJson}', JSON.stringify(pkg, null, 2));"
+        info "    [info]  writing 'game_exe_file' to config file: '$game_exe_name'"
+
+        # check if a '$game_exe_name' is not empty
+        if [[ -n "$game_exe_name" ]]; then
+            node -e "let pkg=require('${toxicConfigJson}'); pkg['exe_path'] = \"ToxicGame\/drive_c\/Games\/${install_folder_name}\/${game_exe_name}\"; require('fs').writeFileSync('${toxicConfigJson}', JSON.stringify(pkg, null, 2));"
+
+            if [[ "$my_global_file_type" == *"mt"* ]]; then
+                node -e "let pkg=require('${toxicConfigJson}'); pkg['backend'] = 'dxmt'; require('fs').writeFileSync('${toxicConfigJson}', JSON.stringify(pkg, null, 2));"
+            fi
+        else
+            error "    [error]  'game_exe_name' is empty!"
         fi
-    else
-        error "    [error]  'game_exe_name' is empty!"
     fi
 }
 
@@ -201,6 +197,10 @@ update_wine_games() {
             write_game_exe_file_to_config_file "$install_folder_name" "$old_version_file_name" "$new_version_file_name" "$old_version_app_path" "$new_version_app_path" "$game_exe_name"
         elif [ -d "$old_version_app_path" ]; then
             do_when_old_file_exist "$install_folder_name" "$old_version_file_name" "$new_version_file_name" "$old_version_app_path" "$new_version_app_path" "$game_exe_name"
+
+            sleep 80
+            do_when_new_file_exist "$install_folder_name" "$old_version_file_name" "$new_version_file_name" "$old_version_app_path" "$new_version_app_path" "$game_exe_name"
+            write_game_exe_file_to_config_file "$install_folder_name" "$old_version_file_name" "$new_version_file_name" "$old_version_app_path" "$new_version_app_path" "$game_exe_name"
         fi
     done
 }
